Sidebar menu (#443)
* Added Sidebar component

* Clear unnecessary comments

* Added sidebar folder

* Edit svg icons

* updated SidebarToggle icon

* Code clean

* Added config API file

* Added activeTab function

* Added Suport svg icon

* Cleared unnecessary <g> tags in svg icons

* Changed some colors value to variables

* Changed position and size of svg icons

* Changed toggle icon style
