+++ src/features/sidebar/index.js
@@ -1,0 +1,1 @@
+export { Sidebar } from './sidebar.js'
